feat: packer.nvim -> lazy.nvim へ変更して plugin 設定を更新
以下の変更も一緒に実施
- 'arcticicestudio/nord-vim' -> 'shaunsingh/nord.nvim'
