+++ nvim/init.lua
@@ -5,20 +5,5 @@ require("display")
 require("plugin")
 
 -- plugin ごとの設定
-require("plugin_nvim-treesitter")
-require("plugin_ale")
-require("plugin_align")
-require("plugin_coc-nvim")
--- require("lsp")
-require("plugin_devicons")
-require("plugin_fzf")
-require("plugin_gina")
 require("plugin_lightline")
--- require('plugin_neogit')
-require("plugin_nerdcommenter")
-require("plugin_nvim-lint")
-require("plugin_vim-terraform")
 require("plugin_lir-nvim")
-require("plugin_null_ls")
-require("plugin_nu")
-require("plugin_silicon")
