--- nvim/lua/plugin-config.lua
@@ -1,2 +1,0 @@
--- coc.nvim
-local opts = { noremap=true, silent=true }
