--- nvim/lua/plugin_align.lua
@@ -1,2 +1,0 @@
--- Alignを日本語環境で使用するための設定
-vim.cmd[[let g:Align_xstrlen = 3]]
