--- nvim/lua/plugin_devicons.lua
@@ -1,2 +1,0 @@
--- フォルダアイコンの表示をON
-vim.cmd[[let g:WebDevIconsUnicodeDecorateFolderNodes = 1]]
