--- nvim/lua/plugin_neogit.lua
@@ -1,2 +1,0 @@
-local neogit = require('neogit')
-neogit.setup {}
