--- nvim/lua/plugin_nerdcommenter.lua
@@ -1,8 +1,0 @@
---  コメントの間にスペースを空ける
-vim.cmd[[let NERDSpaceDelims = 1]]
-
--- <Leader>xでコメントをトグル(NERD_commenter.vim)
--- vim.cmd[[" map <Leader>x, c<space>]]
-
--- 未対応ファイルタイプのエラーメッセージを表示しない
-vim.cmd[[let NERDShutUp=1]]
