--- nvim/lua/plugin_nu.lua
@@ -1,3 +1,0 @@
-require('nu').setup{
-  complete_cmd_names = true
-}
