--- nvim/lua/plugin_silicon.lua
@@ -1,19 +1,0 @@
-vim.cmd[[
-
-let g:silicon = {
-      \   'theme':              'Nord',
-      \   'font':                  'Hack',
-      \   'background':         '#FFFFFF',
-      \   'shadow-color':       '#555555',
-      \   'line-pad':                   2,
-      \   'pad-horiz':                  0,
-      \   'pad-vert':                   0,
-      \   'shadow-blur-radius':         0,
-      \   'shadow-offset-x':            0,
-      \   'shadow-offset-y':            0,
-      \   'line-number':           v:false,
-      \   'round-corner':          v:true,
-      \   'window-controls':       v:true,
-      \ }
-
-]]
