--- nvim/lua/plugin_vim-terraform.lua
@@ -1,1 +1,0 @@
-vim.g.terraform_fmt_on_save=1
